+++ src/idunno.AtProto.Types/RecordKeyFormatException.cs
@@ -21,18 +21,18 @@ namespace idunno.AtProto
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="InvalidResponseTypeException"/> class with a specified error message.
+        /// Initializes a new instance of the <see cref="RecordKeyFormatException"/> class with a specified error message.
         /// </summary>
         public RecordKeyFormatException() : base() { }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="InvalidResponseTypeException"/> class with a specified error message.
+        /// Initializes a new instance of the <see cref="RecordKeyFormatException"/> class with a specified error message.
         /// </summary>
         /// <param name="message">The message that describes the parsing error.</param>
         public RecordKeyFormatException(string message) : base(message) { }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="InvalidResponseTypeException"/> class with a specified error message.
+        /// Initializes a new instance of the <see cref="RecordKeyFormatException"/> class with a specified error message.
         /// </summary>
         /// <param name="message">The message that describes the parsing error.</param>
         /// <param name="inner">The exception that is the cause of the current exception</param>